+++ lexers/moonscript.lua
@@ -12,7 +12,7 @@ lex:add_rule('whitspace', token(lexer.WHITESPACE, lexer.space^1))
 
 -- Table keys.
 lex:add_rule('tbl_key', token('tbl_key', lexer.word * ':' + ':' * lexer.word))
-lex:add_style('tbl_key', lexer.STYLE_REGEX)
+lex:add_style('tbl_key', lexer.styles.regex)
 
 -- Keywords.
 lex:add_rule('keyword', token(lexer.KEYWORD, word_match{